@@ -16,10 +16,10 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1",
+        "php": "^8.2",
         "antaresbr/foundation": "*",
-        "antaresbr/laravel-http": "^10",
-        "laravel/framework": "^10"
+        "antaresbr/laravel-http": "^11",
+        "laravel/framework": "^11"
     },
     "autoload": {
         "files": [
@@ -30,7 +30,7 @@
         }
     },
     "require-dev": {
-        "orchestra/testbench": "^8"
+        "orchestra/testbench": "^9"
     },
     "autoload-dev": {
         "psr-4": {
